backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic==2.5.2
python-dotenv==1.0.0 
boto3==1.28.64
pyotp==2.9.0
redis==5.0.1
